Extract Avrora.Schema.Codec from Avrora.Schema
* Add private client interface tests
