+++ lib/avrora/storage/file.ex
@@ -8,7 +8,7 @@ defmodule Avrora.Storage.File do
 
   require Logger
   alias Avrora.Config
-  alias Avrora.Schema
+  alias Avrora.Schema.Codec, as: SchemaCodec
   alias Avrora.Schema.Name
 
   @behaviour Avrora.Storage
@@ -38,7 +38,7 @@ defmodule Avrora.Storage.File do
   """
   def get(key) when is_binary(key) do
     with {:ok, body} <- read_schema_file_by_name(key),
-         do: Schema.parse(body, &read_schema_file_by_name/1)
+         do: SchemaCodec.from_json(body, &read_schema_file_by_name/1)
   end
 
   def get(key) when is_integer(key), do: {:error, :unsupported}
