+++ test/avrora/codec/plain_test.exs
@@ -183,17 +183,17 @@ defmodule Avrora.Codec.PlainTest do
   defp payment_payload, do: %{"id" => "00000000-0000-0000-0000-000000000000", "amount" => 15.99}
 
   defp payment_schema do
-    {:ok, schema} = Schema.parse(payment_json_schema())
+    {:ok, schema} = Schema.Codec.from_json(payment_json_schema())
     %{schema | id: nil, version: nil}
   end
 
   defp null_value_schema do
-    {:ok, schema} = Schema.parse(null_value_json_schema())
+    {:ok, schema} = Schema.Codec.from_json(null_value_json_schema())
     %{schema | id: nil, version: nil}
   end
 
   defp map_schema do
-    {:ok, schema} = Schema.parse(map_json_schema())
+    {:ok, schema} = Schema.Codec.from_json(map_json_schema())
     %{schema | id: nil, version: nil}
   end
